Allow greater pitch by increasing max UI input value.
For practical reasons
* reduce vertice count in 3D models and
* thus lag

the pitch in general should be allowed to be greater
than 7 despite such threads eventually not being manufactured.

Note that the tip focus is screwed up (maths) for <=M12 screws.
Thus for preset screws even 7 actually is too big. For custom
settings that may not hold true.
Consider Archimedian screws for lifting water.